# backend/docs/swagger.yaml
components:
  schemas:
...
    dto.CreateSpecialty:
      properties:
        title:
          maxLength: 255
          minLength: 1
          type: string
      required:
      - title
      type: object
...
    dto.Error:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
...
    dto.Specialty:
      properties:
        created_at:
          type: string
        id:
          type: string
        title:
          type: string
      type: object
...
    dto.UpdateSpecialty:
      properties:
        title:
          maxLength: 255
          minLength: 1
          type: string
      required:
      - title
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: REST API образовательной системы Edutrack.
  title: Edutrack API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /specialties:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.Specialty'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить список специальностей
      tags:
      - specialties
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateSpecialty'
                description: Данные специальности
                summary: request
        description: Данные специальности
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Specialty'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Создать специальность
      tags:
      - specialties
  /specialties/{id}:
    delete:
      parameters:
      - description: Идентификатор специальности
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Удалить специальность
      tags:
      - specialties
    get:
      parameters:
      - description: Идентификатор специальности
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Specialty'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить специальность по идентификатору
      tags:
      - specialties
    patch:
      parameters:
      - description: Идентификатор специальности
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateSpecialty'
                description: Данные специальности
                summary: request
        description: Данные специальности
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Specialty'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Обновить специальность
      tags:
      - specialties
...
servers:
- url: /api